macos_requirements.txt
mss
numpy
pynput
pyobjc-framework-Quartz
pyyaml
pyautogui